Render new post link directly instead of via filter
